requirements.txt
open-clip-torch
torch
Pillow
transformers
chromadb